feat(styles): add localstorage to themeService
