+++ libs/application/shared/ui/dark-theme/src/lib/dark-theme.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { SetDarkThemeDirective } from './set-dark-theme.directive';
-import { ThemeService } from './theme-service/theme-service.service';
+import { ThemeService } from './theme-service.service';
 
 @NgModule({
   declarations: [
